Separate components into files
