+++ ui/src/AppContent/HomePage/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export function HomePage() {
+  return (
+    <>
+      Nothing here yet
+    </>
+  )
+}
