+++ ui/src/AppFooter/index.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Layout } from 'antd';
+
+const { Footer } = Layout;
+
+export function AppFooter() {
+  const ghSheildImg = "https://img.shields.io/github/stars/btmorr/leifdb?style=social";
+  const ghRepoLink = "https://github.com/btmorr/leifdb";
+  const ghLinkAlt = "View LeifDb on GitHub";
+
+  return (
+    <Footer>
+      ©2020 the LeifDb developers <a href={ghRepoLink}><img src={ghSheildImg} alt={ghLinkAlt} /></a>
+    </Footer>
+  )
+}
